requirements.txt
python-telegram-bot
git+https://github.com/pytgcalls/pytgcalls.git
pyrogram
pytube
apscheduler
asyncio
logging